RevIn in patchtst replaced by commonly used Normalize class
