+++ neuralforecast/_modidx.py
@@ -942,20 +942,6 @@ d = { 'settings': { 'branch': 'main',
                                                                                                               'neuralforecast/models/patchtst.py'),
                                                 'neuralforecast.models.patchtst.PositionalEncoding': ( 'models.patchtst.html#positionalencoding',
                                                                                                        'neuralforecast/models/patchtst.py'),
-                                                'neuralforecast.models.patchtst.RevIN': ( 'models.patchtst.html#revin',
-                                                                                          'neuralforecast/models/patchtst.py'),
-                                                'neuralforecast.models.patchtst.RevIN.__init__': ( 'models.patchtst.html#revin.__init__',
-                                                                                                   'neuralforecast/models/patchtst.py'),
-                                                'neuralforecast.models.patchtst.RevIN._denormalize': ( 'models.patchtst.html#revin._denormalize',
-                                                                                                       'neuralforecast/models/patchtst.py'),
-                                                'neuralforecast.models.patchtst.RevIN._get_statistics': ( 'models.patchtst.html#revin._get_statistics',
-                                                                                                          'neuralforecast/models/patchtst.py'),
-                                                'neuralforecast.models.patchtst.RevIN._init_params': ( 'models.patchtst.html#revin._init_params',
-                                                                                                       'neuralforecast/models/patchtst.py'),
-                                                'neuralforecast.models.patchtst.RevIN._normalize': ( 'models.patchtst.html#revin._normalize',
-                                                                                                     'neuralforecast/models/patchtst.py'),
-                                                'neuralforecast.models.patchtst.RevIN.forward': ( 'models.patchtst.html#revin.forward',
-                                                                                                  'neuralforecast/models/patchtst.py'),
                                                 'neuralforecast.models.patchtst.TSTEncoder': ( 'models.patchtst.html#tstencoder',
                                                                                                'neuralforecast/models/patchtst.py'),
                                                 'neuralforecast.models.patchtst.TSTEncoder.__init__': ( 'models.patchtst.html#tstencoder.__init__',
