+++ neuralforecast/models/patchtst.py
@@ -2,7 +2,7 @@
 
 # %% auto 0
 __all__ = ['SinCosPosEncoding', 'Transpose', 'get_activation_fn', 'PositionalEncoding', 'Coord2dPosEncoding',
-           'Coord1dPosEncoding', 'positional_encoding', 'RevIN', 'PatchTST_backbone', 'Flatten_Head', 'TSTiEncoder',
+           'Coord1dPosEncoding', 'positional_encoding', 'PatchTST_backbone', 'Flatten_Head', 'TSTiEncoder',
            'TSTEncoder', 'TSTEncoderLayer', 'PatchTST']
 
 # %% ../../nbs/models.patchtst.ipynb 5
@@ -15,6 +15,7 @@ import torch.nn as nn
 import torch.nn.functional as F
 
 from ..common._base_windows import BaseWindows
+from ..common._modules import Normalize
 
 from ..losses.pytorch import MAE
 
@@ -134,69 +135,6 @@ def positional_encoding(pe, learn_pe, q_len, hidden_size):
     return nn.Parameter(W_pos, requires_grad=learn_pe)
 
 # %% ../../nbs/models.patchtst.ipynb 13
-class RevIN(nn.Module):
-    def __init__(self, num_features: int, eps=1e-5, affine=True, subtract_last=False):
-        """
-        :param num_features: the number of features or channels
-        :param eps: a value added for numerical stability
-        :param affine: if True, RevIN has learnable affine parameters
-        """
-        super(RevIN, self).__init__()
-        self.num_features = num_features
-        self.eps = eps
-        self.affine = affine
-        self.subtract_last = subtract_last
-        if self.affine:
-            self._init_params()
-
-    def forward(self, x, mode: str):
-        if mode == "norm":
-            self._get_statistics(x)
-            x = self._normalize(x)
-        elif mode == "denorm":
-            x = self._denormalize(x)
-        else:
-            raise NotImplementedError
-        return x
-
-    def _init_params(self):
-        # initialize RevIN params: (C,)
-        self.affine_weight = nn.Parameter(torch.ones(self.num_features))
-        self.affine_bias = nn.Parameter(torch.zeros(self.num_features))
-
-    def _get_statistics(self, x):
-        dim2reduce = tuple(range(1, x.ndim - 1))
-        if self.subtract_last:
-            self.last = x[:, -1, :].unsqueeze(1)
-        else:
-            self.mean = torch.mean(x, dim=dim2reduce, keepdim=True).detach()
-        self.stdev = torch.sqrt(
-            torch.var(x, dim=dim2reduce, keepdim=True, unbiased=False) + self.eps
-        ).detach()
-
-    def _normalize(self, x):
-        if self.subtract_last:
-            x = x - self.last
-        else:
-            x = x - self.mean
-        x = x / self.stdev
-        if self.affine:
-            x = x * self.affine_weight
-            x = x + self.affine_bias
-        return x
-
-    def _denormalize(self, x):
-        if self.affine:
-            x = x - self.affine_bias
-            x = x / (self.affine_weight + self.eps * self.eps)
-        x = x * self.stdev
-        if self.subtract_last:
-            x = x + self.last
-        else:
-            x = x + self.mean
-        return x
-
-# %% ../../nbs/models.patchtst.ipynb 15
 class PatchTST_backbone(nn.Module):
     def __init__(
         self,
@@ -241,7 +179,9 @@ class PatchTST_backbone(nn.Module):
         # RevIn
         self.revin = revin
         if self.revin:
-            self.revin_layer = RevIN(c_in, affine=affine, subtract_last=subtract_last)
+            self.revin_layer = Normalize(
+                c_in, affine=affine, subtract_last=subtract_last
+            )
 
         # Patching
         self.patch_len = patch_len
@@ -818,7 +758,7 @@ class _ScaledDotProductAttention(nn.Module):
         else:
             return output, attn_weights
 
-# %% ../../nbs/models.patchtst.ipynb 17
+# %% ../../nbs/models.patchtst.ipynb 15
 class PatchTST(BaseWindows):
     """PatchTST
